renamed files/restructred routes, to be restful and cleaned up some stuff
