+++ db.js
@@ -11,25 +11,19 @@ var connection = mysql.createConnection({
 connection.connect();
 
 connection.login = function (req, res, next) {
-
     connection.query("SELECT full_name FROM users WHERE email=" + JSON.stringify(req.body.email), function (error, result, field) {
-
         if (error) {
             console.log(error);
         }
         else {
             var json = JSON.parse(JSON.stringify(result));
-            // console.log(json);
             var userName = json[0].full_name;
             req.session.userName = userName;
             req.session.email = req.body.email;
-
             connection.query("SELECT foodName, daysLeft FROM usersFoodData WHERE email=" + JSON.stringify(req.body.email),
-
                 function (error, result, fields) {
                     var usersFood = JSON.parse(JSON.stringify(result));
-                    // console.log(json);
-                    res.render('user', { usersFood: usersFood, userName: userName });
+                    res.render('dashboard', { usersFood: usersFood, userName: userName });
                 });
         }
     });
