+++ public/javascripts/scanner.js
