+++ routes/dashboard.js
@@ -2,42 +2,7 @@ var express = require('express');
 var router = express.Router();
 var db = require('../db');
 
-/****so messy ahhhh****/
-router.get('/login', function(req, res, next){
-  req.body.email = req.session.email;
-  db.login(req, res, next);
-});
-
-/***login user and render dashboard****/
-router.post('/login', function(req, res, next){
-  db.login(req, res, next);
-});
-
-/***register user and render dashboard****/
-router.post('/register', function (req, res, next) {
-
-    var register = {
-    	"full_name": req.body.name,
-    	"email": req.body.email,
-    	"password": req.body.password
-    }
-
-    console.log(register);
-
-    db.query("insert into users set ?", register, function(error) {
-    	if (error) {
-    		console.log(error.message);
-    		throw error;
-    	} else {
-    		console.log("login success");
-    	}
-    });
-    db.login(req, res, next);
-
-
-});
-
-/****ajax response****/
+/****Get all food reference table****/
 router.get('/allFoods', function (req, res, next) {
     /**Pass through foodReference data***/
     var foods;
@@ -51,9 +16,8 @@ router.get('/allFoods', function (req, res, next) {
     });
 });
 
-/*******adding food to fridge ****/
+/*******adding multiple foods to kitchen ****/
 router.post('/addFoodItems', function (req, res, next) {
-
 	// get userID and food info and store them into array.
 	var foods = [];
 	for (var i = 0; i < req.body.foodName.length; i++) {
@@ -62,7 +26,6 @@ router.post('/addFoodItems', function (req, res, next) {
 			req.body.foodName[i],
 			req.body.expiryDate[i]]);
 	};
-
 	// insert food info into database.
 	db.query("insert into usersFoodData (email, foodName, daysLeft) values ?",
 			[foods], function(error) {
@@ -73,19 +36,12 @@ router.post('/addFoodItems', function (req, res, next) {
     		console.log("addition success");
     	}
 	});
-	// redirect user into their dash board.
-  // db.query("SELECT foodName, daysLeft FROM usersFoodData WHERE ID=" + req.session.userID, function (error, result, fields) {
-  //     usersFood = JSON.parse(JSON.stringify(result));
-  //     res.render('user', { usersFood: usersFood, userName: req.session.userName });
-  // });
   req.body.email = req.session.email;
   db.login(req, res, next);
-
 });
 
 /****delete item from fridge****/
 router.delete("/:foodID", function (req, res, next) {
-    
 });
 
 module.exports = router;
