+++ views/about.ejs
@@ -19,7 +19,7 @@
         <div class="collapse navbar-collapse" id="navMenu">
             <ul class="navbar-nav mr-auto">
                 <li class="nav-item active">
-                    <a class="nav-link" href="../index">Home <span class="sr-only">(current)</span></a>
+                    <a class="nav-link" href="../home">Home <span class="sr-only">(current)</span></a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="../about">Features/About</a>
@@ -28,15 +28,15 @@
 
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                    <a id="userPageAnchor" class="nav-link" href="./user/login">Item List</a>
+                    <a id="userPageAnchor" class="nav-link" href="./login">Item List</a>
                 </li>
 
                 <li class="nav-item">
-                    <a id="scanPageAnchor" class="nav-link" href="./user/login">Take A Photo</a>
+                    <a id="scanPageAnchor" class="nav-link" href="./login">Take A Photo</a>
                 </li>
 
                 <li class="nav-item">
-                    <a id="addPageAnchor" class="nav-link" href="./user/login">Add item</a>
+                    <a id="addPageAnchor" class="nav-link" href="./login">Add item</a>
                 </li>
 
             </ul>
@@ -143,7 +143,6 @@ src="https://code.jquery.com/jquery-3.3.1.min.js"
 integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
 crossorigin="anonymous"></script>
 <script type="text/javascript" src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/js/bootstrap.min.js"></script>
-<!-- <script type="text/javascript" src="../javascripts/user.js"></script> -->
 <!--Scripts-->
 </body>
 </html>
