+++ views/dashboard.ejs
@@ -4,9 +4,9 @@
 
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title><%= userName %></title>
+<title><%= userName %> - Dashboard</title>
 
-<link type="text/css" rel="stylesheet" href="../stylesheets/user.css">
+<link type="text/css" rel="stylesheet" href="../stylesheets/dashboard.css">
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css" integrity="sha384-9gVQ4dYFwwWSjIDZnLEWnxCjeSWFphJiwGPXr1jddIhOegiu1FwO5qRGvFXOdJZ4" crossorigin="anonymous">
 
 
@@ -14,7 +14,7 @@
 
 <body>
     <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-        <a class="navbar-brand" href="#">YAMS</a>
+        <a class="navbar-brand" href="./">YAMS</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navMenu" aria-controls="navMenu" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
         </button>
@@ -22,7 +22,7 @@
         <div class="collapse navbar-collapse" id="navMenu">
             <ul class="navbar-nav mr-auto">
                 <li class="nav-item active">
-                    <a class="nav-link" href="../index">Home <span class="sr-only">(current)</span></a>
+                    <a class="nav-link" href="../home">Home <span class="sr-only">(current)</span></a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="../about">Features/About</a>
@@ -41,7 +41,7 @@
                 <li class="nav-item">
                     <a id="addPageAnchor" class="nav-link" href="#">Add item</a>
                 </li>
-                
+
             </ul>
 
         </div>
@@ -162,8 +162,8 @@ integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
 crossorigin="anonymous"></script>
 <script src="https://unpkg.com/popper.js/dist/umd/popper.min.js"></script>
 <script type="text/javascript" src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/js/bootstrap.min.js"></script>
-<script type="text/javascript" src="../javascripts/user.js"></script>
-<script type="text/javascript" src="../javascripts/scan.js"></script>
+<script type="text/javascript" src="../javascripts/dashboard.js"></script>
+<script type="text/javascript" src="../javascripts/scanner.js"></script>
 
 </body>
 </html>
